Allow steps in docker build script to be skipped using environment variables
Skip default-judgehost-chroot image by default because it's only
needed for dj_make_chroot_docker.

Skip help message by default because it is no longer accurate (some of
the images might not exist, and the instructions for pushing to Docker
Hub aren't relevant).

+++ docker/build.sh
@@ -15,7 +15,7 @@ fi
 
 VERSION="$1"
 
-if false; then
+if [ "${BUILD_DOWNLOAD_RELEASE-0}" != 0 ]; then
 URL=https://www.domjudge.org/releases/domjudge-${VERSION}.tar.gz
 FILE=domjudge.tar.gz
 
@@ -30,25 +30,34 @@ fi
 echo "[ok] DOMjudge version ${VERSION} downloaded as domjudge.tar.gz"; echo
 fi
 
+if [ "${BUILD_FROM_SOURCE-1}" != 0 ]; then
 echo "[..] Boostrapping..."
 tar c -C .. --exclude=./.git --exclude="./docker/*.tar.gz" . |
 	docker build --target=dist -t domjudge/dist:${VERSION} -f docker/domserver/Dockerfile.source -
 docker run --rm domjudge/dist:${VERSION} tar cz -C /domjudge-src domjudge > domjudge.tar.gz
 docker rmi --no-prune domjudge/dist:${VERSION}
 echo "[ok] Done boostrapping"
+fi
 
+if [ "${BUILD_DOMSERVER-1}" != 0 ]; then
 echo "[..] Building Docker image for domserver..."
 ./build-domjudge.sh "domjudge/domserver:${VERSION}"
 echo "[ok] Done building Docker image for domserver"
+fi
 
+if [ "${BUILD_JUDGEHOST-1}" != 0 ]; then
 echo "[..] Building Docker image for judgehost using intermediate build image..."
 ./build-judgehost.sh "domjudge/judgehost:${VERSION}"
 echo "[ok] Done building Docker image for judgehost"
+fi
 
+if [ "${BUILD_DEFAULT_JUDGEHOST_CHROOT-0}" != 0 ]; then
 echo "[..] Building Docker image for judgehost chroot..."
 docker build -t "domjudge/default-judgehost-chroot:${VERSION}" -f judgehost/Dockerfile.chroot .
 echo "[ok] Done building Docker image for judgehost chroot"
+fi
 
+if [ "${BUILD_PRINT_HELP-0}" != 0 ]; then
 echo "All done. Image domjudge/domserver:${VERSION} and domjudge/judgehost:${VERSION} created"
 echo "If you are a DOMjudge maintainer with access to the domjudge organization on Docker Hub, you can now run the following command to push them to Docker Hub:"
 echo "$ docker push domjudge/domserver:${VERSION} && docker push domjudge/judgehost:${VERSION} && docker push domjudge/default-judgehost-chroot:${VERSION}"
@@ -57,3 +66,4 @@ echo "$ docker tag domjudge/domserver:${VERSION} domjudge/domserver:latest && \
 docker tag domjudge/judgehost:${VERSION} domjudge/judgehost:latest && \
 docker tag domjudge/default-judgehost-chroot:${VERSION} domjudge/default-judgehost-chroot:latest && \
 docker push domjudge/domserver:latest && docker push domjudge/judgehost:latest && docker push domjudge/default-judgehost-chroot:latest"
+fi
